pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "goated"
version = "0.1.0"
description = "Live Kalshi commodity theo engine — Pyth-anchored, numba hot path, <200ms full book"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.11",
    "numba>=0.59",
    "pyyaml>=6.0",
    "websockets>=12.0",
    "httpx>=0.27",
    "cryptography>=42.0",
    "structlog>=24.1",
    "python-dateutil>=2.9",
    "pytz>=2024.1",
    "duckdb>=1.0",
    "ib_insync>=0.9.86",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-benchmark>=4.0",
    "hypothesis>=6.100",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
include = ["feeds*", "state*", "models*", "engine*", "calibration*", "validation*", "trufaidp*", "hedge*"]
exclude = ["tests*", "benchmarks*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q"

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "PL"]
ignore = ["PLR0913", "PLR2004"]
